Team_ALL-IS-WELL_SURYAPRAKASH_AJAYKUMAR_Basic_Calculator.cpp
The aim is to build a simple calculator application that offers various arithmetic operations and additional features. The calculator should be capable of performing basic arithmetic operations like addition, subtraction, multiplication, and division for both integers and floating-point numbers. Furthermore, it should support exponentiation and determine the number of factors of a given integer.